Add alembic and stop deleting/recreating all tables.  (#22)
Add alembic to api
Generate first alembic migration creating all tables
Add `init_db()` which runs migrations if postgres, or drops and recreates all tables if sqlite
Run `init_db()` on fastapi server startup 

Misc:
Change imports to be absolute 
Remove Dockerfile.eve
Move where .env.api is expected to be found

+++ apps/api/src/db/setup.py
@@ -5,7 +5,8 @@ from sqlalchemy import URL
 from sqlmodel import Session as SQLModelSession
 from sqlmodel import SQLModel, create_engine
 
-from .models import *  # noqa
+# Don't need to import models because alembic will populate metadata
+# from .models import *  # noqa
 
 db_password = os.getenv("POSTGRES_DB_PASSWORD")
 db_host = os.getenv("POSTGRES_DB_HOST")
@@ -20,13 +21,7 @@ if not is_test and (db_password and db_host):
     )
     connect_args = {}
 else:
-    SQLALCHEMY_DATABASE_URL = URL.create(
-        drivername="sqlite",
-        username="",
-        password="",
-        host="",
-        database="test.db",
-    )
+    SQLALCHEMY_DATABASE_URL = URL.create(drivername="sqlite", database="./test.db")
 
     connect_args = {"check_same_thread": False}
 
@@ -42,7 +37,20 @@ def Session():
         session.close()
 
 
-# TODO: Don't do this in prod
-with Session() as session:
-    SQLModel.metadata.drop_all(session.get_bind())
-    SQLModel.metadata.create_all(session.get_bind())
+def init_db():
+    """
+    Initializes database tables
+    If sqlite, then it drops and creates tables
+    If postgres, then it runs alembic migrations
+    """
+    if SQLALCHEMY_DATABASE_URL.drivername == "sqlite":
+        with Session() as session:
+            SQLModel.metadata.drop_all(session.get_bind())
+            SQLModel.metadata.create_all(session.get_bind())
+    else:
+        from alembic import command
+        from alembic.config import Config
+
+        dir_path = os.path.dirname(os.path.realpath(__file__))
+        alembic_cfg = Config(os.path.join(dir_path, "../../alembic.ini"))
+        command.upgrade(alembic_cfg, "head")
